LPD-30131 Move folders
