+++ modules/test/playwright/tests/workspaces/liferay-workspace-marketplace/pages/customer-dashboard/customerDashboardAppDetailsPage.ts
